samples/python/agents/pydantic_ai/pydantic_ai_a2a_demo/requirements.txt
pydantic-ai-slim[a2a]
uvicorn
fastapi
httpx
python-dotenv 